pull out BaseLoadStateView
